app/requirements.txt
aiogram>=3.13.1
aiohttp>=3.9.1
aiofiles>=23.2.0
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
sqlalchemy[asyncio]>=2.0.23
aiosqlite>=0.20.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
alembic>=1.13.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
cryptography>=41.0.8
python-dotenv>=1.0.0
pydantic>=2.5.0
python-dateutil>=2.8.2
pytz>=2023.3
pytest>=7.4.3
pytest-asyncio>=0.21.1
httpx>=0.25.2
faker>=20.1.0
black>=23.11.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.7.1